pyproject.toml
[project]
name = "fund-lens-etl"
version = "0.1.0"
description = ""
authors = [
    {name = "Tom Boone",email = "tbc@tomboone.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "alembic (>=1.17.1,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "prefect (>=3.4.25,<4.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ipykernel = "^7.0.1"
notebook = "^7.4.7"
nbstripout = "^0.8.1"
pre-commit = "^4.3.0"
ruff = "^0.14.2"
bandit = "^1.8.6"
mypy = "^1.18.2"
types-requests = "^2.32.4.20250913"
types-python-dateutil = "^2.9.0.20251008"
isort = "^7.0.0"
black = "^25.9.0"
pandas-stubs = "^2.3.2.250926"

[tool.ruff]
line-length = 100  # Increase from 88 to reduce line-too-long errors

[tool.ruff.lint]  # Move under 'lint' section
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # Ignore line-too-long errors

[tool.ruff.lint.isort]  # Move under 'lint.isort'
known-first-party = ["fund_lens_etl"]
